refactor(test): move unit tests for Subject in new file
